+++ Tests/FluentAssertions.Specs/Equivalency/DataEquivalencySpecs.DataColumn.cs
@@ -47,7 +47,7 @@ namespace FluentAssertions.Specs
 
                 // Assert
                 action.Should().Throw<XunitException>().WithMessage(
-                    "Expected *of type DataColumn*to be non-null, but found null*");
+                    "Expected *to be non-null, but found null*");
             }
 
             [Fact]