+++ Tests/FluentAssertions.Specs/Equivalency/DataEquivalencySpecs.DataSet.Typed.cs
@@ -45,7 +45,7 @@ namespace FluentAssertions.Specs
 
                 // Assert
                 action.Should().Throw<XunitException>().WithMessage(
-                    "Expected *of type DataSet*to be non-null, but found null*");
+                    "Expected *to be non-null, but found null*");
             }
 
             [Fact]